dummyproject/board/requirements.txt
django
djangorestframework
django-filter
djangocorsheaders